Reflecting changes of library source code in the ccextractor source code
